Bug 571147 extend 'lic.net' with common facilities
Extract `api` package to emulate unnecessary `net.api` bundle
