+++ bundles/org.eclipse.passage.lbc.base/src/org/eclipse/passage/lbc/internal/base/EObjectTransfer.java
@@ -16,7 +16,7 @@ import org.eclipse.emf.ecore.EObject;
 import org.eclipse.passage.lic.internal.api.LicensingException;
 import org.eclipse.passage.lic.internal.api.conditions.mining.ContentType;
 import org.eclipse.passage.lic.internal.emf.EObjectToBytes;
-import org.eclipse.passage.lic.internal.net.handle.NetResponse;
+import org.eclipse.passage.lic.internal.net.api.handle.NetResponse;
 
 public final class EObjectTransfer implements NetResponse {
